Alter our RPATH adjustment behavior.
Introduce new method adjust_rpath() which does the following:

    - Lists the existing RPATH value in the .so.  We can't make any
      assumptions about where this path may be rooted -- with R, it
      will almost always live outside of $PREFIX/lib.

    - Finds the common root path between the existing RPATH and our
      $PREFIX/lib directory.  If the common root *is* $PREFIX/lib, we
      can set a relative $ORIGIN/../.. RPATH, with the number of ../'s
      determined by depth of .so directory.

    - If we can't trace back a common root path to $PREFIX/lib, we need
      to have an RPATH with two directories; one that is relative to the
      directory the target .so lives in, the other being relative to
      $PREFIX/lib.  Without the latter, things like libgfortran.so will end
      up linking against /usr/lib libs first, not our $PREFIX/lib libs.

There are a number of packages that exercise this functionality (in that if
this functionality isn't present, the packages aren't generated properly and
can't either be linked a) at all, or b) in a relocatable fashion).

Most package recipes live in github.com/tpn/cran-conda-recipes.

    - r
    - r-rcpp
    - r-rcppeigen
    - jags
    - r-rjags

If you can load RccpEigen and Rjags in R (via `library(RccpEigen)` and
`library(Rjags)`), especially if R has been installed into a separate
environment, that is an excellent litmus test for all the RPATH behavior
working correctly.

Note: with all that being said, I'm still not 100% sure this change addresses
everything.  There are a few erroneous `ldd` outputs I'm seeing with this
change in place.